remove unnecessary export_test
