+++ go/cs/reservation/segment/BUILD.bazel
@@ -21,7 +21,6 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "export_test.go",
         "path_test.go",
         "reservation_test.go",
     ],
