+++ go/cs/reservation/segment/path_test.go
@@ -17,6 +17,8 @@ package segment
 import (
 	"testing"
 
+	"github.com/scionproto/scion/go/lib/common"
+	"github.com/scionproto/scion/go/lib/xtest"
 	"github.com/stretchr/testify/require"
 )
 
@@ -97,4 +99,19 @@ func TestEqualPath(t *testing.T) {
 			require.Equal(t, tc.IsEqual, eq)
 		})
 	}
+}
+
+func newPathFromComponents(chain ...interface{}) Path {
+	if len(chain)%3 != 0 {
+		panic("wrong number of arguments")
+	}
+	p := Path{}
+	for i := 0; i < len(chain); i += 3 {
+		p = append(p, PathStep{
+			Ingress: common.IFIDType(chain[i].(int)),
+			AS:      xtest.MustParseAS(chain[i+1].(string)),
+			Egress:  common.IFIDType(chain[i+2].(int)),
+		})
+	}
+	return p
 }
